+++ inst/apps/GreatPlainsBCGCalc/global.R
@@ -1,7 +1,7 @@
 # Shiny Global File
 
 # Version ----
-pkg_version <- "0.0.3.9016"
+pkg_version <- "0.0.3.9021"
 
 # Packages----
 library(BCGcalc)
@@ -25,6 +25,7 @@ library(rmarkdown)
 library(tools)
 library(openxlsx)
 library(leaflet)
+library(kableExtra) # reports
 
 # Source ----
 
@@ -93,7 +94,12 @@ dn_files_ref    <- "reference"
 dn_files_input  <- "_user_input"
 
 # Selection Choices----
-sel_community <- c("Fish", "Bugs_IA", "Bugs_KS", "Bugs_MO", "Bugs_NE")
+sel_community <- c("Fish",
+                   "Bugs_IA",
+                   "Bugs_KS",
+                   "Bugs_MO",
+                   "Bugs_NE",
+                   "Bugs_OK")
 
 ##  BCG Models ----
 url_bcg_base <- "https://github.com/leppott/BCGcalc/raw/main/inst/extdata"